refactor(tui): extract FileItem into widgets/file_item.rs
Move FileItem struct, FileItem::new(), impl From<SessionInfo>, and
format_size() helper into a dedicated file_item module. Migrate related
tests (format_size_works, file_item_from_session_info) to the new file.
