+++ src/tui/widgets/file_explorer.rs
@@ -12,7 +12,6 @@
 use std::collections::HashSet;
 use std::path::Path;
 
-use chrono::{DateTime, Local};
 use ratatui::{
     buffer::Buffer,
     layout::{Constraint, Layout, Rect},
@@ -23,62 +22,9 @@ use ratatui::{
 
 use crate::asciicast::EventType;
 use crate::files::backup::has_backup;
-use crate::storage::SessionInfo;
 use crate::theme::current_theme;
 
-/// A file item in the explorer
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct FileItem {
-    /// Full path to the file
-    pub path: String,
-    /// Display name (filename without path)
-    pub name: String,
-    /// Agent name (e.g., "claude", "codex")
-    pub agent: String,
-    /// File size in bytes
-    pub size: u64,
-    /// Last modified time
-    pub modified: DateTime<Local>,
-    /// Whether a backup file exists for this item (cached)
-    pub has_backup: bool,
-}
-
-impl FileItem {
-    /// Create a new FileItem
-    pub fn new(
-        path: impl Into<String>,
-        name: impl Into<String>,
-        agent: impl Into<String>,
-        size: u64,
-        modified: DateTime<Local>,
-    ) -> Self {
-        let path_str = path.into();
-        let has_backup = has_backup(std::path::Path::new(&path_str));
-        Self {
-            path: path_str,
-            name: name.into(),
-            agent: agent.into(),
-            size,
-            modified,
-            has_backup,
-        }
-    }
-}
-
-impl From<SessionInfo> for FileItem {
-    fn from(session: SessionInfo) -> Self {
-        let path_str = session.path.to_string_lossy().to_string();
-        let has_backup = has_backup(std::path::Path::new(&path_str));
-        Self {
-            path: path_str,
-            name: session.filename,
-            agent: session.agent,
-            size: session.size,
-            modified: session.modified,
-            has_backup,
-        }
-    }
-}
+use super::file_item::{format_size, FileItem};
 
 use crate::terminal::{Color, StyledLine};
 
@@ -1034,27 +980,10 @@ impl Widget for FileExplorerWidget<'_> {
     }
 }
 
-/// Format a byte size as human-readable string
-fn format_size(bytes: u64) -> String {
-    const KB: u64 = 1024;
-    const MB: u64 = KB * 1024;
-    const GB: u64 = MB * 1024;
-
-    if bytes >= GB {
-        format!("{:.1} GB", bytes as f64 / GB as f64)
-    } else if bytes >= MB {
-        format!("{:.1} MB", bytes as f64 / MB as f64)
-    } else if bytes >= KB {
-        format!("{:.1} KB", bytes as f64 / KB as f64)
-    } else {
-        format!("{} B", bytes)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use chrono::TimeZone;
+    use chrono::{Local, TimeZone};
 
     fn create_test_items() -> Vec<FileItem> {
         vec![
@@ -1315,15 +1244,6 @@ mod tests {
         // No panic means success
     }
 
-    #[test]
-    fn format_size_works() {
-        assert_eq!(format_size(500), "500 B");
-        assert_eq!(format_size(1024), "1.0 KB");
-        assert_eq!(format_size(1536), "1.5 KB");
-        assert_eq!(format_size(1048576), "1.0 MB");
-        assert_eq!(format_size(1073741824), "1.0 GB");
-    }
-
     // Search filter tests
 
     #[test]
@@ -1397,29 +1317,6 @@ mod tests {
         assert!(explorer.search_filter().is_none());
         explorer.set_search_filter(Some("test".to_string()));
         assert_eq!(explorer.search_filter(), Some("test"));
-    }
-
-    // From<SessionInfo> conversion test
-    #[test]
-    fn file_item_from_session_info() {
-        use std::path::PathBuf;
-
-        let session = SessionInfo {
-            path: PathBuf::from("/sessions/claude/test.cast"),
-            agent: "claude".to_string(),
-            filename: "test.cast".to_string(),
-            size: 1024,
-            modified: Local.with_ymd_and_hms(2024, 1, 15, 10, 0, 0).unwrap(),
-            age_days: 0,
-            age_hours: 0,
-            age_minutes: 0,
-        };
-
-        let item = FileItem::from(session);
-        assert_eq!(item.path, "/sessions/claude/test.cast");
-        assert_eq!(item.name, "test.cast");
-        assert_eq!(item.agent, "claude");
-        assert_eq!(item.size, 1024);
     }
 
     // SessionPreview tests
