+++ src/tui/widgets/mod.rs
@@ -3,9 +3,11 @@
 //! Reusable UI components for the terminal interface.
 
 pub mod file_explorer;
+pub mod file_item;
 pub mod logo;
 
 pub use file_explorer::{
-    FileExplorer, FileExplorerWidget, FileItem, SessionPreview, SortDirection, SortField,
+    FileExplorer, FileExplorerWidget, SessionPreview, SortDirection, SortField,
 };
+pub use file_item::{format_size, FileItem};
 pub use logo::Logo;
